+++ src/firmware/shared/math/polynomial_1d.c
