+++ src/firmware/shared/math/vector_2d.h
