+++ src/shared/BUILD
@@ -11,7 +11,7 @@ cc_library(
     name = "robot_constants",
     hdrs = ["robot_constants.h"],
     deps = [
-        "//firmware/shared:tbots_math",
+        "//firmware/shared/math:tbots_math",
     ],
 )
